refactor: extract frame_compare CLI wiring into cli_entry
